deduplicate loadData&sampleSpec into train.examples.cifar10
