+++ client/benchmark/src/main/java/org/eu/fedcampus/benchmark/BenchmarkCifar10Worker.kt
@@ -3,10 +3,11 @@ package org.eu.fedcampus.benchmark
 import android.content.Context
 import android.util.Log
 import androidx.work.WorkerParameters
-import org.eu.fedcampus.train.SampleSpec
 import org.eu.fedcampus.train.background.BaseTrainWorker
-import org.eu.fedcampus.train.helpers.classifierAccuracy
-import org.eu.fedcampus.train.helpers.negativeLogLikelihoodLoss
+import org.eu.fedcampus.train.examples.cifar10.DATA_TYPE
+import org.eu.fedcampus.train.examples.cifar10.Float3DArray
+import org.eu.fedcampus.train.examples.cifar10.loadData
+import org.eu.fedcampus.train.examples.cifar10.sampleSpec
 
 class BenchmarkCifar10Worker(context: Context, params: WorkerParameters) :
     BaseTrainWorker<Float3DArray, FloatArray>(
@@ -25,16 +26,3 @@ class BenchmarkCifar10Worker(context: Context, params: WorkerParameters) :
 fun logTrain(msg: String) {
     Log.i(BenchmarkCifar10Worker.TAG, msg)
 }
-
-fun sampleSpec() = SampleSpec<Float3DArray, FloatArray>(
-    { it.toTypedArray() },
-    { it.toTypedArray() },
-    { Array(it) { FloatArray(10) } },
-    ::negativeLogLikelihoodLoss,
-    ::classifierAccuracy,
-)
-
-
-const val DATA_TYPE = "CIFAR10_32x32x3"
-
-typealias Float3DArray = Array<Array<FloatArray>>
